server/pyproject.toml
[tool.poetry]
name = "nabazon_server"
version = "0.1.0"
description = ""
authors = ["peleg68 <pelegtsa@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
starlette_prometheus = "^0.5.0"
pyyaml = "^5.3.1"
poetry = "^1.0.5"
loguru = "^0.4.1"
uvicorn = "^0.11.3"
pydantic = "^1.4"
fastapi = "^0.53.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-cov = "^2.8.1"

[tool.poetry.scripts]
app = 'nabazon_server.main:main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

